feat(form): align form to menu
